Functions

GL2017Replication.ModelGL — Type

Structure holding all parameters and solutions to the steady-state model

GL2017Replication.TransGL — Type

Structure holding transition objects

GL2017Replication.EGM! — Method

EGM!(gl::ModelGL)

Calculates optimal policy functions with the EGM algorithm.

GL2017Replication.EGM_trans! — Method

EGM_trans!(gl::ModelGL,Tgl::TransGL,t::Int)

Calculates optimal policy functions with one step of the EGM algorithm.

GL2017Replication.aggregate! — Method

aggregate!(gl::ModelGL;terminal::Bool = false)

Computes aggregate values and their distance from target. Requires that the joint distribution has already been solved (compute_distribution!) If ´terminal = true´ only the terminal debt target will be considered.

GL2017Replication.calibrate! — Method

calibrate!(gl::ModelGL)

Calibrate the model to the target values. Does not require any prior function calls. If load_params = true, the calibrated parameters are preloaded to speed up the calibration.

GL2017Replication.calibrate_terminal — Method

gl_tss = calibrate_terminal(gl_initial::ModelGL)

Calibrate the model to the terminal target value (only borrowing constraint) Requires a calibrated initital steady-state object. Returns a new terminal steady-state object. If load_params = true, the calibrated parameters are preloaded to speed up the calibration.

GL2017Replication.compute_distribution! — Method

compute_distribution!(gl::ModelGL)

Computes the joint distribution of productivity and bond holdings. Requires that optimal policies have already been solved (EGM!)

GL2017Replication.compute_steady_state! — Method

compute_steady_state!(gl::ModelGL)

Computes steady-state policy functions with given set of parameters.

GL2017Replication.describe — Method

    describe(gl::ModelGL,gl_2::ModelGL)

Print solution status as well as values and description of parameters and aggregate values: Comparison of initial and terminal steady-state.

GL2017Replication.describe — Method

    describe(gl::ModelGL)

Print solution status as well as values and description of parameters and aggregate values.

GL2017Replication.domath — Method

domath(x::Number)

Return x + 5.

GL2017Replication.find_cl — Method

F = find_cl(c, j, b1, b2, r, θ, z, fac, gameta)

Helper function to find consumption at the constraint. Returns the residual the budget constraint.

c - consumption

j - state index

b1 - bond holdings today

b2 - bond holdings tomorrow

r - interest rate

θ - state value array

fac - parameter

gameta - parameter

Return

F - residual budget constraint

GL2017Replication.hello — Method

hello(who::String)

Return "Hello, who".

GL2017Replication.initilize! — Method

initilize!(gl::ModelGL)

Adds borrowing constraint as gridpoint

GL2017Replication.load_parameters_iss! — Method

load_parameters_iss!(gl))

Pre-load calibrated parameters to speed up calibration.

GL2017Replication.load_parameters_tss! — Method

load_parameters_tss!(gl))

Pre-load calibrated parameters to speed up calibration (terminal).

GL2017Replication.plots_figure_1 — Method

    (F1,p1,p2) = plots_figure_1(gl::ModelGL)

Plots figure I graphs.

GL2017Replication.plots_figure_3 — Method

    (F3,p1,p2,p3,p4)  = plots_figure_4(gl::ModelGL,gl_tss::ModelGL)

Plots figure III graphs.

GL2017Replication.plots_figure_4 — Method

    (F4,p1,p2) = plots_figure_4(gl::ModelGL,gl_tss::ModelGL)

Plots figure IV graphs.

GL2017Replication.print_params — Method

    print_params(gl::ModelGL,gl_2::ModelGL)

Print values and description of parameters and aggregate values: Comparison of initial and terminal steady-state.

GL2017Replication.print_params — Method

    print_params(gl::ModelGL)

Print values and description of parameters and aggregate values.

GL2017Replication.print_status — Method

    print_status(gl::ModelGL,gl_2::ModelGL)

Print solution status of both models.

GL2017Replication.print_status — Method

    print_status(gl::ModelGL)

Print solution status.

GL2017Replication.transition! — Method

transition!(gl::ModelGL,gl_tss::ModelGL,Tgl::TransGL)

Computes transition path.

end